删除无用资源

+++ app/src/main/java/com/example/practice/customview/picker/view/PickerScrollView.java
@@ -220,8 +220,8 @@ public class PickerScrollView extends View {
      * @param type     1表示向下绘制，-1表示向上绘制
      */
     private void drawOtherText(Canvas canvas, int position, int type) {
-        float d = (float) (MARGIN_ALPHA * mMinTextSize * position + type
-                * mMoveLen);
+        float d = MARGIN_ALPHA * mMinTextSize * position + type
+                * mMoveLen;
         float scale = parabola(mViewHeight / 4.0f, d);
         float size = (mMaxTextSize - mMinTextSize) * scale + mMinTextSize;
         mPaint.setTextSize(size);
